stop the detector before listeners go out of scope
Fixes SDK-3841

Previously the detector was being stopped in run(), AFTER the appropriate process_XXX_input method was called.  However, when using FrameDetector, this caused intermittent crashes at shutdown time because the listener objects are defined in those process_XXX_input methods, and therefore they go out of scope when those methods return.  Since the background processing thread in the SDK may still be processing frames at that point and invoking the listeners,  that's bad.

The fix is to move the stop() call into the process_XXX_input methods so that background processing finishes before the listeners go out of scope.